+++ pkg/config/config.go
@@ -2,7 +2,7 @@ package config
 
 import (
 	"errors"
-	"log"
+	"fmt"
 	"os"
 	"regexp"
 	"strings"
@@ -40,6 +40,7 @@ type LocationConfig struct {
 }
 
 type EmployeeConfig struct {
+	EmployeeID   string `yaml:"employee_ID,omitempty"`
 	Department   string `yaml:"department,omitempty"`
 	JobTitle     string `yaml:"job_title,omitempty"`
 	Type         string `yaml:"type,omitempty"`
@@ -57,6 +58,7 @@ type GroupConfig struct {
 	WhoCanPostMessage    string         `yaml:"who_can_post,omitempty"`
 	WhoCanJoin           string         `yaml:"who_can_join,omitempty"`
 	AllowExternalMembers string         `yaml:"allow_external_members,omitempty"`
+	IsArchived           string         `yaml:"is_archived,omitempty"`
 	Members              []MemberConfig `yaml:"members,omitempty"`
 }
 
@@ -103,34 +105,72 @@ func SaveToFile(config *Config, filename string) error {
 	return nil
 }
 
-func (c *Config) Validate() error {
+func (c *Config) Validate() []error {
+	var allTheErrors []error
+	re := regexp.MustCompile("^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$")
+	re164 := regexp.MustCompile("^\\+[1-9]\\d{1,14}$")
+
 	// validate organization
 	if c.Organization == "" {
-		return errors.New("no organization configured")
+		//return errors.New("no organization configured")
+		allTheErrors = append(allTheErrors, errors.New("no organization configured"))
 	}
-
-	re := regexp.MustCompile("^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$")
 
 	//validate users
 	userEmails := []string{}
 	for _, user := range c.Users {
 		if util.StringSliceContains(userEmails, user.PrimaryEmail) {
-			log.Fatalf("Validation failed: duplicate user defined (user: %s)", user.PrimaryEmail)
+			//return fmt.Errorf("duplicate user defined (user: %s)", user.PrimaryEmail)
+			allTheErrors = append(allTheErrors, fmt.Errorf("duplicate user defined (user: %s)", user.PrimaryEmail))
 		}
 
 		if user.PrimaryEmail == user.SecondaryEmail {
-			log.Fatalf("Validation failed: user has defined the same primary and secondary email (user: %s)", user.PrimaryEmail)
+			//return fmt.Errorf("user has defined the same primary and secondary email (user: %s)", user.PrimaryEmail)
+			allTheErrors = append(allTheErrors, fmt.Errorf("user has defined the same primary and secondary email (user: %s)", user.PrimaryEmail))
 		}
 
 		if re.MatchString(user.PrimaryEmail) == false {
-			log.Fatalf("Validation failed: invalid primary email (user: %s)", user.PrimaryEmail)
+			//return fmt.Errorf("invalid format of primary email (user: %s)", user.PrimaryEmail)
+			allTheErrors = append(allTheErrors, fmt.Errorf("primary email is not a valid email-address (user: %s)", user.PrimaryEmail))
 		}
 
 		if user.SecondaryEmail != "" {
 			if re.MatchString(user.SecondaryEmail) == false {
-				log.Fatalf("Validation failed: invalid secondary email (user: %s)", user.PrimaryEmail)
+				//return fmt.Errorf("invalid format of secondary email (user: %s)", user.PrimaryEmail)
+				allTheErrors = append(allTheErrors, fmt.Errorf("secondary email is not a valid email-address (user: %s)", user.PrimaryEmail))
 			}
 		}
+
+		if user.RecoveryEmail != "" {
+			if re.MatchString(user.RecoveryEmail) == false {
+				//return fmt.Errorf("invalid format of recovery email (user: %s)", user.PrimaryEmail)
+				allTheErrors = append(allTheErrors, fmt.Errorf("recovery email is not a valid email-address (user: %s)", user.PrimaryEmail))
+			}
+		}
+
+		if len(user.Aliases) > 0 {
+			for _, alias := range user.Aliases {
+				if re.MatchString(alias) == false {
+					//return fmt.Errorf("invalid format of alias email (user: %s)", user.PrimaryEmail)
+					allTheErrors = append(allTheErrors, fmt.Errorf("alias email is not a valid email-address (user: %s)", user.PrimaryEmail))
+				}
+			}
+		}
+
+		if user.Employee.ManagerEmail != "" {
+			if re.MatchString(user.Employee.ManagerEmail) == false {
+				//return fmt.Errorf("invalid format of the manager's email (user: %s)", user.PrimaryEmail)
+				allTheErrors = append(allTheErrors, fmt.Errorf("manager's email is not a valid email-address (user: %s)", user.PrimaryEmail))
+			}
+		}
+
+		if user.RecoveryPhone != "" {
+			if re164.MatchString(user.RecoveryPhone) == false {
+				//return fmt.Errorf("invalid format of recovery phone (user: %s). The phone number must be in the E.164 format, starting with the plus sign (+). Example: +16506661212.", user.PrimaryEmail)
+				allTheErrors = append(allTheErrors, fmt.Errorf("invalid format of recovery phone (user: %s). The phone number must be in the E.164 format, starting with the plus sign (+). Example: +16506661212.", user.PrimaryEmail))
+			}
+		}
+
 		userEmails = append(userEmails, user.PrimaryEmail)
 	}
 
@@ -138,21 +178,55 @@ func (c *Config) Validate() error {
 	groupEmails := []string{}
 	for _, group := range c.Groups {
 		if util.StringSliceContains(groupEmails, group.Email) {
-			log.Fatalf("Validation failed: duplicate group email defined (%s)", group.Email)
+			//return fmt.Errorf("duplicate group email defined (%s)", group.Email)
+			allTheErrors = append(allTheErrors, fmt.Errorf("duplicate group email defined (%s)", group.Email))
 		}
 
 		if re.MatchString(group.Email) == false {
-			log.Fatalf("Validation failed: invalid group email (%s)", group.Email)
+			//return fmt.Errorf("invalid group email (%s)", group.Email)
+			allTheErrors = append(allTheErrors, fmt.Errorf("group email is not a valid email-address (%s)", group.Email))
+		}
+
+		if !(strings.EqualFold(group.WhoCanContactOwner, "ALL_IN_DOMAIN_CAN_CONTACT") || strings.EqualFold(group.WhoCanContactOwner, "ALL_MANAGERS_CAN_CONTACT") || strings.EqualFold(group.WhoCanContactOwner, "ALL_MEMBERS_CAN_CONTACT") || strings.EqualFold(group.WhoCanContactOwner, "ANYONE_CAN_CONTACT")) {
+			//return fmt.Errorf("wrong value specified for 'who_can_contact_owner' field (group: %s)", group.Name)
+			allTheErrors = append(allTheErrors, fmt.Errorf("wrong value specified for 'who_can_contact_owner' field (group: %s)", group.Name))
+		}
+
+		if !(strings.EqualFold(group.WhoCanViewMembership, "ALL_IN_DOMAIN_CAN_VIEW") || strings.EqualFold(group.WhoCanViewMembership, "ALL_MEMBERS_CAN_VIEW") || strings.EqualFold(group.WhoCanViewMembership, "ALL_MANAGERS_CAN_VIEW")) {
+			//return fmt.Errorf("wrong value specified for 'who_can_view_members' field (group: %s)", group.Name)
+			allTheErrors = append(allTheErrors, fmt.Errorf("wrong value specified for 'who_can_view_members' field (group: %s)", group.Name))
+		}
+
+		if !(strings.EqualFold(group.WhoCanApproveMembers, "ALL_MEMBERS_CAN_APPROVE") || strings.EqualFold(group.WhoCanApproveMembers, "ALL_MANAGERS_CAN_APPROVE") || strings.EqualFold(group.WhoCanApproveMembers, "ALL_OWNERS_CAN_APPROVE") || strings.EqualFold(group.WhoCanApproveMembers, "NONE_CAN_APPROVE")) {
+			//return fmt.Errorf("wrong value specified for 'who_can_approve_members' field (group: %s)", group.Name)
+			allTheErrors = append(allTheErrors, fmt.Errorf("wrong value specified for 'who_can_approve_members' field (group: %s)", group.Name))
+		}
+
+		if !(strings.EqualFold(group.WhoCanPostMessage, "NONE_CAN_POST") || strings.EqualFold(group.WhoCanPostMessage, "ALL_MANAGERS_CAN_POST") || strings.EqualFold(group.WhoCanPostMessage, "ALL_MEMBERS_CAN_POST") || strings.EqualFold(group.WhoCanPostMessage, "ALL_OWNERS_CAN_POST") || strings.EqualFold(group.WhoCanPostMessage, "ALL_IN_DOMAIN_CAN_POST") || strings.EqualFold(group.WhoCanPostMessage, "ANYONE_CAN_POST")) {
+			//return fmt.Errorf("wrong value specified for 'who_can_post' field (group: %s)", group.Name)
+			allTheErrors = append(allTheErrors, fmt.Errorf("wrong value specified for 'who_can_post' field (group: %s)", group.Name))
+		}
+
+		if !(strings.EqualFold(group.WhoCanJoin, "CAN_REQUEST_TO_JOIN") || strings.EqualFold(group.WhoCanJoin, "INVITED_CAN_JOIN") || strings.EqualFold(group.WhoCanJoin, "ALL_IN_DOMAIN_CAN_JOIN") || strings.EqualFold(group.WhoCanJoin, "ANYONE_CAN_JOIN")) {
+			//return fmt.Errorf("wrong value specified for 'who_can_contact_owner' field (group: %s)", group.Name)
+			allTheErrors = append(allTheErrors, fmt.Errorf("wrong value specified for 'who_can_contact_owner' field (group: %s)", group.Name))
+		}
+
+		if !(strings.EqualFold(group.AllowExternalMembers, "true") || strings.EqualFold(group.AllowExternalMembers, "false")) {
+			//return fmt.Errorf("wrong value specified for 'allow_external_members' field (group: %s)", group.Name)
+			allTheErrors = append(allTheErrors, fmt.Errorf("wrong value specified for 'allow_external_members' field (group: %s)", group.Name))
 		}
 
 		memberEmails := []string{}
 		for _, member := range group.Members {
 			if util.StringSliceContains(memberEmails, member.Email) {
-				log.Fatalf("Validation failed: duplicate member defined in a group (group: %s, member: %s)", group.Name, member.Email)
+				//return fmt.Errorf("duplicate member defined in a group (group: %s, member: %s)", group.Name, member.Email)
+				allTheErrors = append(allTheErrors, fmt.Errorf("duplicate member defined in a group (group: %s, member: %s)", group.Name, member.Email))
 			}
 
 			if !(strings.EqualFold(member.Role, "OWNER") || strings.EqualFold(member.Role, "MANAGER") || strings.EqualFold(member.Role, "MEMBER")) {
-				log.Fatalf("Validation failed: wrong member role specified (group: %s, member: %s)", group.Name, member.Email)
+				//return fmt.Errorf("wrong member role specified (group: %s, member: %s)", group.Name, member.Email)
+				allTheErrors = append(allTheErrors, fmt.Errorf("wrong member role specified (group: %s, member: %s). Permitted values are OWNER, MEMBER or MANAGER.", group.Name, member.Email))
 			}
 		}
 	}
@@ -161,16 +235,24 @@ func (c *Config) Validate() error {
 	ouNames := []string{}
 	for _, ou := range c.OrgUnits {
 		if util.StringSliceContains(ouNames, ou.Name) {
-			log.Fatalf("Validation failed: duplicate org unit defined (%s)", ou.Name)
+			//return fmt.Errorf("duplicate org unit defined (%s)", ou.Name)
+			allTheErrors = append(allTheErrors, fmt.Errorf("duplicate org unit defined (%s)", ou.Name))
 		}
 
 		if ou.ParentOrgUnitPath[0] != '/' {
-			log.Fatalf("Validation failed: wrong ParentOrgUnitPath specified for org unit (%s)", ou.Name)
+			//return fmt.Errorf("wrong ParentOrgUnitPath specified for org unit (%s)", ou.Name)
+			allTheErrors = append(allTheErrors, fmt.Errorf("wrong ParentOrgUnitPath specified for org unit (%s)", ou.Name))
+
 		}
 		if ou.OrgUnitPath[0] != '/' {
-			log.Fatalf("Validation failed: wrong OrgUnitPath specified for org unit (%s)", ou.Name)
+			//return fmt.Errorf("wrong OrgUnitPath specified for org unit (%s)", ou.Name)
+			allTheErrors = append(allTheErrors, fmt.Errorf("wrong OrgUnitPath specified for org unit (%s)", ou.Name))
 		}
 	}
 
-	return nil
+	if allTheErrors != nil {
+		return allTheErrors
+	} else {
+		return nil
+	}
 }