+++ pkg/glib/glib.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"fmt"
 	"io/ioutil"
-	"log"
 
 	"github.com/kubermatic-labs/gman/pkg/config"
 	password "github.com/sethvargo/go-password/password"
@@ -22,12 +21,12 @@ func NewDirectoryService(clientSecretFile string, impersonatedUserEmail string) 
 
 	jsonCredentials, err := ioutil.ReadFile(clientSecretFile)
 	if err != nil {
-		return nil, fmt.Errorf("ReadFile(clientSecretFile): %v", err)
+		return nil, fmt.Errorf("Unable to read json credentials (clientSecretFile): %v", err)
 	}
 
 	config, err := google.JWTConfigFromJSON(jsonCredentials, admin.AdminDirectoryUserScope, admin.AdminDirectoryGroupScope, admin.AdminDirectoryGroupMemberScope, admin.AdminDirectoryOrgunitScope, admin.AdminDirectoryResourceCalendarScope)
 	if err != nil {
-		return nil, fmt.Errorf("JWTConfigFromJSON: %v", err)
+		return nil, fmt.Errorf("Unable to process credentials: %v", err)
 	}
 	config.Subject = impersonatedUserEmail
 
@@ -35,7 +34,7 @@ func NewDirectoryService(clientSecretFile string, impersonatedUserEmail string) 
 
 	srv, err := admin.NewService(ctx, option.WithTokenSource(ts))
 	if err != nil {
-		return nil, fmt.Errorf("NewService: %v", err)
+		return nil, fmt.Errorf("Unable to create a new Admin Service: %v", err)
 	}
 	return srv, nil
 }
@@ -47,12 +46,12 @@ func NewGroupsService(clientSecretFile string, impersonatedUserEmail string) (*g
 
 	jsonCredentials, err := ioutil.ReadFile(clientSecretFile)
 	if err != nil {
-		return nil, fmt.Errorf("ReadFile(clientSecretFile): %v", err)
+		return nil, fmt.Errorf("Unable to read json credentials (clientSecretFile): %v", err)
 	}
 
 	config, err := google.JWTConfigFromJSON(jsonCredentials, groupssettings.AppsGroupsSettingsScope)
 	if err != nil {
-		return nil, fmt.Errorf("JWTConfigFromJSON: %v", err)
+		return nil, fmt.Errorf("Unable to process credentials: %v", err)
 	}
 	config.Subject = impersonatedUserEmail
 
@@ -60,7 +59,7 @@ func NewGroupsService(clientSecretFile string, impersonatedUserEmail string) (*g
 
 	srv, err := groupssettings.NewService(ctx, option.WithTokenSource(ts))
 	if err != nil {
-		return nil, fmt.Errorf("NewService: %v", err)
+		return nil, fmt.Errorf("Unable to create a new Groupssettings Service: %v", err)
 	}
 	return srv, nil
 }
@@ -73,10 +72,8 @@ func NewGroupsService(clientSecretFile string, impersonatedUserEmail string) (*g
 func GetListOfUsers(srv admin.Service) ([]*admin.User, error) {
 	request, err := srv.Users.List().Customer("my_customer").OrderBy("email").Do()
 	if err != nil {
-		log.Fatalf("Unable to retrieve users in domain: %v", err)
-		return nil, err
+		return nil, fmt.Errorf("Unable to retrieve list of users in domain: %v", err)
 	}
-
 	return request.Users, nil
 }
 
@@ -100,8 +97,7 @@ func CreateUser(srv admin.Service, user *config.UserConfig) error {
 	// generate a rand password
 	pass, err := password.Generate(20, 5, 5, false, false)
 	if err != nil {
-		log.Fatalf("Unable to generate password: %v", err)
-		return err
+		return fmt.Errorf("Unable to generate password: %v", err)
 	}
 	newUser := createGSuiteUserFromConfig(srv, user)
 	newUser.Password = pass
@@ -109,9 +105,11 @@ func CreateUser(srv admin.Service, user *config.UserConfig) error {
 
 	_, err = srv.Users.Insert(newUser).Do()
 	if err != nil {
-		log.Fatalf("Unable to create a user: %v", err)
-		return err
+		return fmt.Errorf("Unable to insert a user: %v", err)
 	}
+
+	HandleUserAliases(srv, newUser, user.Aliases)
+
 	return nil
 }
 
@@ -119,8 +117,7 @@ func CreateUser(srv admin.Service, user *config.UserConfig) error {
 func DeleteUser(srv admin.Service, user *admin.User) error {
 	err := srv.Users.Delete(user.PrimaryEmail).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete a user: %v", err)
-		return err
+		return fmt.Errorf("Unable to delete a user %s: %v", user.PrimaryEmail, err)
 	}
 	return nil
 }
@@ -130,9 +127,69 @@ func UpdateUser(srv admin.Service, user *config.UserConfig) error {
 	updatedUser := createGSuiteUserFromConfig(srv, user)
 	_, err := srv.Users.Update(user.PrimaryEmail, updatedUser).Do()
 	if err != nil {
-		log.Fatalf("Unable to update a user: %v", err)
-		return err
+		return fmt.Errorf("Unable to update a user %s: %v", user.PrimaryEmail, err)
 	}
+
+	HandleUserAliases(srv, updatedUser, user.Aliases)
+
+	return nil
+}
+
+func HandleUserAliases(srv admin.Service, googleUser *admin.User, configAliases []string) error {
+	request, err := srv.Users.Aliases.List(googleUser.PrimaryEmail).Do()
+	if err != nil {
+		return fmt.Errorf("Unable to list user aliases in GSuite: %v", err)
+	}
+
+	if len(configAliases) == 0 {
+		for _, alias := range request.Aliases {
+			err = srv.Users.Aliases.Delete(googleUser.PrimaryEmail, fmt.Sprint(alias.(map[string]interface{})["alias"])).Do()
+			if err != nil {
+				return fmt.Errorf("Unable to delete user alias: %v", err)
+			}
+		}
+	} else {
+		// check aliases to delete
+		for _, alias := range request.Aliases {
+			found := false
+			for _, configAlias := range configAliases {
+				if alias.(map[string]interface{})["alias"] == configAlias {
+					found = true
+					break
+				}
+			}
+			if !found {
+				// delete
+				err = srv.Users.Aliases.Delete(googleUser.PrimaryEmail, fmt.Sprint(alias.(map[string]interface{})["alias"])).Do()
+				if err != nil {
+					return fmt.Errorf("Unable to delete user alias: %v", err)
+				}
+			}
+
+		}
+	}
+
+	// check aliases to add
+	for _, configAlias := range configAliases {
+		found := false
+		for _, alias := range request.Aliases {
+			if alias.(map[string]interface{})["alias"] == configAlias {
+				found = true
+				break
+			}
+		}
+		if !found {
+			// add
+			newAlias := &admin.Alias{
+				Alias: configAlias,
+			}
+			_, err = srv.Users.Aliases.Insert(googleUser.PrimaryEmail, newAlias).Do()
+			if err != nil {
+				return fmt.Errorf("Unable to add user alias: %v", err)
+			}
+		}
+	}
+
 	return nil
 }
 
@@ -145,13 +202,6 @@ func createGSuiteUserFromConfig(srv admin.Service, user *config.UserConfig) *adm
 		},
 		PrimaryEmail: user.PrimaryEmail,
 		OrgUnitPath:  user.OrgUnitPath,
-	}
-
-	if len(user.Aliases) > 0 {
-		// TODO: fix it, doesnt add aliases
-		for _, alias := range user.Aliases {
-			googleUser.Aliases = append(googleUser.Aliases, string(alias))
-		}
 	}
 
 	if len(user.Phones) > 0 {
@@ -167,7 +217,7 @@ func createGSuiteUserFromConfig(srv admin.Service, user *config.UserConfig) *adm
 	}
 
 	if user.Address != "" {
-		addr := &[]admin.UserAddress{
+		addr := []admin.UserAddress{
 			{
 				Formatted: user.Address,
 				Type:      "home",
@@ -185,7 +235,7 @@ func createGSuiteUserFromConfig(srv admin.Service, user *config.UserConfig) *adm
 	}
 
 	if user.SecondaryEmail != "" {
-		workEm := &[]admin.UserEmail{
+		workEm := []admin.UserEmail{
 			{
 				Address: user.SecondaryEmail,
 				Type:    "work",
@@ -195,7 +245,7 @@ func createGSuiteUserFromConfig(srv admin.Service, user *config.UserConfig) *adm
 	}
 
 	if user.Employee != (config.EmployeeConfig{}) {
-		uOrg := &[]admin.UserOrganization{
+		uOrg := []admin.UserOrganization{
 			{
 				Department:  user.Employee.Department,
 				Title:       user.Employee.JobTitle,
@@ -204,7 +254,7 @@ func createGSuiteUserFromConfig(srv admin.Service, user *config.UserConfig) *adm
 			},
 		}
 		googleUser.Organizations = uOrg
-		rel := &[]admin.UserRelation{
+		rel := []admin.UserRelation{
 			{
 				Value: user.Employee.ManagerEmail,
 				Type:  "manager",
@@ -212,20 +262,27 @@ func createGSuiteUserFromConfig(srv admin.Service, user *config.UserConfig) *adm
 		}
 		googleUser.Relations = rel
 
+		ids := []admin.UserExternalId{
+			{
+				Value: user.Employee.EmployeeID,
+				Type:  "organization",
+			},
+		}
+		googleUser.ExternalIds = ids
+
 	}
 
 	if user.Location != (config.LocationConfig{}) {
-		//BuildingHandling(srv, user.Location)
-		loc := &[]admin.UserLocation{
+		loc := []admin.UserLocation{
 			{
+				Area:         "desk",
 				BuildingId:   user.Location.Building,
 				FloorName:    user.Location.Floor,
 				FloorSection: user.Location.FloorSection,
 				Type:         "desk",
 			},
 		}
-		//googleUser.Locations = loc
-		fmt.Println(loc)
+		googleUser.Locations = loc
 
 	}
 
@@ -256,6 +313,14 @@ func CreateConfigUserFromGSuite(googleUser *admin.User) config.UserConfig {
 	if googleUser.Phones != nil {
 		for _, phone := range googleUser.Phones.([]interface{}) {
 			configUser.Phones = append(configUser.Phones, fmt.Sprint(phone.(map[string]interface{})["value"]))
+		}
+	}
+
+	if googleUser.ExternalIds != nil {
+		for _, id := range googleUser.ExternalIds.([]interface{}) {
+			if id.(map[string]interface{})["type"] == "organization" {
+				configUser.Employee.EmployeeID = fmt.Sprint(id.(map[string]interface{})["value"])
+			}
 		}
 	}
 
@@ -302,8 +367,7 @@ func CreateConfigUserFromGSuite(googleUser *admin.User) config.UserConfig {
 func GetListOfGroups(srv *admin.Service) ([]*admin.Group, error) {
 	request, err := srv.Groups.List().Customer("my_customer").OrderBy("email").Do()
 	if err != nil {
-		log.Fatalf("Unable to retrieve groups in domain: %v", err)
-		return nil, err
+		return nil, fmt.Errorf("Unable to retrieve a list of groups in domain: %v", err)
 	}
 	return request.Groups, nil
 }
@@ -312,26 +376,27 @@ func GetListOfGroups(srv *admin.Service) ([]*admin.Group, error) {
 func GetSettingOfGroup(srv *groupssettings.Service, groupId string) (*groupssettings.Groups, error) {
 	request, err := srv.Groups.Get(groupId).Do()
 	if err != nil {
-		log.Fatalf("Unable to retrieve group settings: %v", err)
-		return nil, err
+		return nil, fmt.Errorf("Unable to retrieve group's (%s) settings: %v", groupId, err)
 	}
 	return request, nil
 }
 
 // CreateGroup creates a new group in GSuite via their API
-func CreateGroup(srv admin.Service, group *config.GroupConfig) error {
-	newGroup, groupSettings := createGSuiteGroupFromConfig(group)
+func CreateGroup(srv admin.Service, grSrv groupssettings.Service, group *config.GroupConfig) error {
+	newGroup, groupSettings := CreateGSuiteGroupFromConfig(group)
 	_, err := srv.Groups.Insert(newGroup).Do()
 	if err != nil {
-		log.Fatalf("Unable to create a group: %v", err)
-		return err
+		return fmt.Errorf("Unable to insert a group: %v", err)
 	}
 	// add the members
 	for _, member := range group.Members {
 		AddNewMember(srv, newGroup.Email, &member)
 	}
-	fmt.Println(groupSettings)
-
+	// add the group's settings
+	_, err = grSrv.Groups.Update(newGroup.Email, groupSettings).Do()
+	if err != nil {
+		return fmt.Errorf("Unable to set the group settings: %v", err)
+	}
 	return nil
 }
 
@@ -339,26 +404,29 @@ func CreateGroup(srv admin.Service, group *config.GroupConfig) error {
 func DeleteGroup(srv admin.Service, group *admin.Group) error {
 	err := srv.Groups.Delete(group.Email).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete a group: %v", err)
-		return err
+		return fmt.Errorf("Unable to delete a group: %v", err)
 	}
 	return nil
 }
 
 // UpdateGroup updates the remote group with config
-func UpdateGroup(srv admin.Service, group *config.GroupConfig) error {
-	updatedGroup, groupSettings := createGSuiteGroupFromConfig(group)
+func UpdateGroup(srv admin.Service, grSrv groupssettings.Service, group *config.GroupConfig) error {
+	updatedGroup, groupSettings := CreateGSuiteGroupFromConfig(group)
 	_, err := srv.Groups.Update(group.Email, updatedGroup).Do()
 	if err != nil {
-		log.Fatalf("Unable to update a group: %v", err)
-		return err
+		return fmt.Errorf("Unable to update a group: %v", err)
 	}
-	fmt.Println(groupSettings)
+	// update group's settings
+	_, err = grSrv.Groups.Update(group.Email, groupSettings).Do()
+	if err != nil {
+		return fmt.Errorf("Unable to update group settings: %v", err)
+	}
+
 	return nil
 }
 
 // createGSuiteGroupFromConfig converts a ConfigGroup to (GSuite) admin.Group
-func createGSuiteGroupFromConfig(group *config.GroupConfig) (*admin.Group, *groupssettings.Groups) {
+func CreateGSuiteGroupFromConfig(group *config.GroupConfig) (*admin.Group, *groupssettings.Groups) {
 	googleGroup := &admin.Group{
 		Name:  group.Name,
 		Email: group.Email,
@@ -373,10 +441,38 @@ func createGSuiteGroupFromConfig(group *config.GroupConfig) (*admin.Group, *grou
 		WhoCanApproveMembers: group.WhoCanApproveMembers,
 		WhoCanPostMessage:    group.WhoCanPostMessage,
 		WhoCanJoin:           group.WhoCanJoin,
+		IsArchived:           group.IsArchived,
+		ArchiveOnly:          group.IsArchived,
 		AllowExternalMembers: group.AllowExternalMembers,
 	}
 
 	return googleGroup, groupSettings
+}
+
+func CreateConfigGroupFromGSuite(googleGroup *admin.Group, members []*admin.Member, gSettings *groupssettings.Groups) config.GroupConfig {
+	configGroup := config.GroupConfig{
+		Name:                 googleGroup.Name,
+		Email:                googleGroup.Email,
+		Description:          googleGroup.Description,
+		WhoCanContactOwner:   gSettings.WhoCanContactOwner,
+		WhoCanViewMembership: gSettings.WhoCanViewMembership,
+		WhoCanApproveMembers: gSettings.WhoCanApproveMembers,
+		WhoCanPostMessage:    gSettings.WhoCanPostMessage,
+		WhoCanJoin:           gSettings.WhoCanJoin,
+		AllowExternalMembers: gSettings.AllowExternalMembers,
+		IsArchived:           gSettings.IsArchived,
+		Members:              []config.MemberConfig{},
+	}
+
+	for _, m := range members {
+		configGroup.Members = append(configGroup.Members, config.MemberConfig{
+			Email: m.Email,
+			Role:  m.Role,
+		})
+
+	}
+
+	return configGroup
 }
 
 //----------------------------------------//
@@ -387,8 +483,7 @@ func createGSuiteGroupFromConfig(group *config.GroupConfig) (*admin.Group, *grou
 func GetListOfMembers(srv *admin.Service, group *admin.Group) ([]*admin.Member, error) {
 	request, err := srv.Members.List(group.Email).Do()
 	if err != nil {
-		log.Fatalf("Unable to retrieve members in group %s: %v", group.Name, err)
-		return nil, err
+		return nil, fmt.Errorf("Unable to retrieve members in group %s: %v", group.Name, err)
 	}
 	return request.Members, nil
 }
@@ -398,8 +493,7 @@ func AddNewMember(srv admin.Service, groupEmail string, member *config.MemberCon
 	newMember := createGSuiteGroupMemberFromConfig(member)
 	_, err := srv.Members.Insert(groupEmail, newMember).Do()
 	if err != nil {
-		log.Fatalf("Unable to add a member: %v", err)
-		return err
+		return fmt.Errorf("Unable to add a member to a group: %v", err)
 	}
 	return nil
 }
@@ -408,20 +502,18 @@ func AddNewMember(srv admin.Service, groupEmail string, member *config.MemberCon
 func RemoveMember(srv admin.Service, groupEmail string, member *admin.Member) error {
 	err := srv.Members.Delete(groupEmail, member.Email).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete a member: %v", err)
-		return err
+		return fmt.Errorf("Unable to delete a member from a group: %v", err)
 	}
 	return nil
 }
 
 // MemberExists checks if member exists in group
-func MemberExists(srv admin.Service, group *admin.Group, member *config.MemberConfig) bool {
+func MemberExists(srv admin.Service, group *admin.Group, member *config.MemberConfig) (bool, error) {
 	exists, err := srv.Members.HasMember(group.Email, member.Email).Do()
 	if err != nil {
-		log.Fatalf("Unable to check if member %s exists in a group %s: %v", member.Email, group.Name, err)
-		return false
+		return false, fmt.Errorf("Unable to check if member %s exists in a group %s: %v", member.Email, group.Name, err)
 	}
-	return exists.IsMember
+	return exists.IsMember, nil
 }
 
 // UpdateMembership changes the role of the member
@@ -430,8 +522,7 @@ func UpdateMembership(srv admin.Service, groupEmail string, member *config.Membe
 	newMember := createGSuiteGroupMemberFromConfig(member)
 	_, err := srv.Members.Update(groupEmail, member.Email, newMember).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete a member: %v", err)
-		return err
+		return fmt.Errorf("Unable to update a member in a group: %v", err)
 	}
 	return nil
 }
@@ -453,8 +544,7 @@ func createGSuiteGroupMemberFromConfig(member *config.MemberConfig) *admin.Membe
 func GetListOfOrgUnits(srv *admin.Service) ([]*admin.OrgUnit, error) {
 	request, err := srv.Orgunits.List("my_customer").Type("all").Do()
 	if err != nil {
-		log.Fatalf("Unable to retrieve OrgUnits in domain: %v", err)
-		return nil, err
+		return nil, fmt.Errorf("Unable to list OrgUnits in domain: %v", err)
 	}
 	return request.OrganizationUnits, nil
 }
@@ -464,8 +554,7 @@ func CreateOrgUnit(srv admin.Service, ou *config.OrgUnitConfig) error {
 	newOU := createGSuiteOUFromConfig(ou)
 	_, err := srv.Orgunits.Insert("my_customer", newOU).Do()
 	if err != nil {
-		log.Fatalf("Unable to create an org unit: %v", err)
-		return err
+		return fmt.Errorf("Unable to create an org unit: %v", err)
 	}
 	return nil
 }
@@ -482,8 +571,7 @@ func DeleteOrgUnit(srv admin.Service, ou *admin.OrgUnit) error {
 
 	err := srv.Orgunits.Delete("my_customer", orgUPath).Do()
 	if err != nil {
-		log.Fatalf("Unable to delete an org unit: %v", err)
-		return err
+		return fmt.Errorf("Unable to delete an org unit: %v", err)
 	}
 	return nil
 }
@@ -501,8 +589,7 @@ func UpdateOrgUnit(srv admin.Service, ou *config.OrgUnitConfig) error {
 
 	_, err := srv.Orgunits.Update("my_customer", orgUPath, updatedOu).Do()
 	if err != nil {
-		log.Fatalf("Unable to update an org unit: %v", err)
-		return err
+		return fmt.Errorf("Unable to update an org unit: %v", err)
 	}
 	return nil
 }
@@ -519,44 +606,4 @@ func createGSuiteOUFromConfig(ou *config.OrgUnitConfig) *admin.OrgUnit {
 	}
 
 	return googleOU
-}
-
-func BuildingHandling(srv admin.Service, location config.LocationConfig) error {
-	request, err := srv.Resources.Buildings.List("my_customer").Do()
-	if err != nil {
-		log.Fatalf("Unable to retrieve Buildings in domain: %v", err)
-		return err
-	}
-
-	found := false
-	for _, building := range request.Buildings {
-		fmt.Println(building)
-		if location.Building == building.BuildingId {
-			found = true
-			// update with floor & floor section
-			//for _, floor := range building.FloorNames {
-
-			//}
-
-			break
-		}
-	}
-	if !found {
-		// (r *ResourcesBuildingsService) Insert(customer string, building *Building)
-		newBuilding := &admin.Building{
-			BuildingId:   location.Building,
-			BuildingName: location.Building,
-			FloorNames:   []string{location.Floor},
-			//FloorSection: location.FloorSection,
-			//ResourceName: fmt.Sprintf("%s, %s, %s", location.Building, location.Floor, location.FloorSection),
-			//ResourceId:   location.Building,
-		}
-		_, err := srv.Resources.Buildings.Insert("my_customer", newBuilding).Do()
-		if err != nil {
-			log.Fatalf("Unable to create Building in domain: %v", err)
-			return err
-		}
-	}
-
-	return nil
 }