+++ pkg/sync/sync.go
@@ -20,7 +20,7 @@ func SyncConfiguration(ctx context.Context, cfg *config.Config, clientService *a
 	if err := SyncUsers(ctx, clientService, cfg, confirm); err != nil {
 		return fmt.Errorf("failed to sync users: %v", err)
 	}
-	if err := SyncGroups(ctx, clientService, cfg, confirm); err != nil {
+	if err := SyncGroups(ctx, clientService, groupService, cfg, confirm); err != nil {
 		return fmt.Errorf("failed to sync groups: %v", err)
 	}
 
@@ -39,7 +39,7 @@ func SyncUsers(ctx context.Context, clientService *admin.Service, cfg *config.Co
 	// get the current users array
 	currentUsers, err := glib.GetListOfUsers(*clientService)
 	if err != nil {
-		return fmt.Errorf("⚠ failed to get current users: %v", err)
+		return err
 	}
 	// config defined users
 	configUsers := cfg.Users
@@ -56,7 +56,6 @@ func SyncUsers(ctx context.Context, clientService *admin.Service, cfg *config.Co
 					// user is existing & should exist, so check if needs an update
 					currentUserConfig := glib.CreateConfigUserFromGSuite(currentUser)
 					if !reflect.DeepEqual(currentUserConfig, configUser) {
-						fmt.Println(currentUser.Locations)
 						usersToUpdate = append(usersToUpdate, configUser)
 					}
 					break
@@ -85,22 +84,34 @@ func SyncUsers(ctx context.Context, clientService *admin.Service, cfg *config.Co
 		if usersToCreate != nil {
 			log.Println("✎ Creating...")
 			for _, user := range usersToCreate {
-				glib.CreateUser(*clientService, &user)
-				log.Printf("\t+ user: %s\n", user.PrimaryEmail)
+				err := glib.CreateUser(*clientService, &user)
+				if err != nil {
+					return fmt.Errorf("⚠ Failed to create user %s: %v.", user.PrimaryEmail, err)
+				} else {
+					log.Printf("\t+ user: %s\n", user.PrimaryEmail)
+				}
 			}
 		}
 		if usersToDelete != nil {
 			log.Println("✁ Deleting...")
 			for _, user := range usersToDelete {
-				glib.DeleteUser(*clientService, user)
-				log.Printf("\t- user: %s\n", user.PrimaryEmail)
+				err := glib.DeleteUser(*clientService, user)
+				if err != nil {
+					return fmt.Errorf("⚠ Failed to delete user %s: %v.", user.PrimaryEmail, err)
+				} else {
+					log.Printf("\t- user: %s\n", user.PrimaryEmail)
+				}
 			}
 		}
 		if usersToUpdate != nil {
 			log.Println("✎ Updating...")
 			for _, user := range usersToUpdate {
-				glib.UpdateUser(*clientService, &user)
-				log.Printf("\t~ user: %s\n", user.PrimaryEmail)
+				err := glib.UpdateUser(*clientService, &user)
+				if err != nil {
+					return fmt.Errorf("⚠ Failed to update user %s: %v.", user.PrimaryEmail, err)
+				} else {
+					log.Printf("\t~ user: %s\n", user.PrimaryEmail)
+				}
 			}
 		}
 	} else {
@@ -146,7 +157,7 @@ type groupUpdate struct {
 }
 
 // SyncGroups
-func SyncGroups(ctx context.Context, clientService *admin.Service, cfg *config.Config, confirm bool) error {
+func SyncGroups(ctx context.Context, clientService *admin.Service, groupService *groupssettings.Service, cfg *config.Config, confirm bool) error {
 	var (
 		groupsToDelete []*admin.Group
 		groupsToCreate []config.GroupConfig
@@ -156,9 +167,8 @@ func SyncGroups(ctx context.Context, clientService *admin.Service, cfg *config.C
 	log.Println("⇄ Syncing groups")
 	// get the current groups array
 	currentGroups, err := glib.GetListOfGroups(clientService)
-
 	if err != nil {
-		return fmt.Errorf("⚠ failed to get current groups: %v", err)
+		return err
 	}
 	// config defined groups
 	configGroups := cfg.Groups
@@ -175,10 +185,16 @@ func SyncGroups(ctx context.Context, clientService *admin.Service, cfg *config.C
 					// group is existing & should exist, so check if needs an update
 					var upGroup groupUpdate
 					upGroup.membersToAdd, upGroup.membersToRemove, upGroup.membersToUpdate = SyncMembers(ctx, clientService, &cfgGroup, currGroup)
-					if cfgGroup.Name != currGroup.Name ||
-						cfgGroup.Description != currGroup.Description ||
-						upGroup.membersToAdd != nil || upGroup.membersToRemove != nil ||
-						upGroup.membersToUpdate != nil {
+					currentMembers, err := glib.GetListOfMembers(clientService, currGroup)
+					if err != nil {
+						return err
+					}
+					currentSettings, err := glib.GetSettingOfGroup(groupService, currGroup.Email)
+					if err != nil {
+						return err
+					}
+					currentGroupConfig := glib.CreateConfigGroupFromGSuite(currGroup, currentMembers, currentSettings)
+					if !reflect.DeepEqual(currentGroupConfig, cfgGroup) {
 						upGroup.groupToUpdate = cfgGroup
 						groupsToUpdate = append(groupsToUpdate, upGroup)
 					}
@@ -210,36 +226,58 @@ func SyncGroups(ctx context.Context, clientService *admin.Service, cfg *config.C
 		if groupsToCreate != nil {
 			log.Println("✎ Creating...")
 			for _, gr := range groupsToCreate {
-				glib.CreateGroup(*clientService, &gr)
-				log.Printf("\t+ group: %s\n", gr.Name)
+				err := glib.CreateGroup(*clientService, *groupService, &gr)
+				if err != nil {
+					return fmt.Errorf("⚠ Failed to create a group %s: %v.", gr.Name, err)
+				} else {
+					log.Printf("\t+ group: %s\n", gr.Name)
+				}
 			}
 		}
 		if groupsToDelete != nil {
 			log.Println("✁ Deleting...")
 			for _, gr := range groupsToDelete {
-				glib.DeleteGroup(*clientService, gr)
-				log.Printf("\t- group: %s\n", gr.Name)
+				err := glib.DeleteGroup(*clientService, gr)
+				if err != nil {
+					return fmt.Errorf("⚠ Failed to delete a group %s: %v.", gr.Name, err)
+				} else {
+					log.Printf("\t- group: %s\n", gr.Name)
+				}
 			}
 		}
 		if groupsToUpdate != nil {
 			log.Println("✎ Updating...")
 			for _, gr := range groupsToUpdate {
-				glib.UpdateGroup(*clientService, &gr.groupToUpdate)
-				log.Printf("\t~ group: %s\n", gr.groupToUpdate.Name)
+				err := glib.UpdateGroup(*clientService, *groupService, &gr.groupToUpdate)
+				if err != nil {
+					return fmt.Errorf("⚠ Failed to update a group: %v.", err)
+				} else {
+					log.Printf("\t~ group: %s\n", gr.groupToUpdate.Name)
+				}
 
 				for _, mem := range gr.membersToAdd {
-					log.Printf("\t\t+ %s \n", mem.Email)
-					glib.AddNewMember(*clientService, gr.groupToUpdate.Email, mem)
-
+					err := glib.AddNewMember(*clientService, gr.groupToUpdate.Email, mem)
+					if err != nil {
+						return fmt.Errorf("⚠ Failed to add a member to a group: %v.", err)
+					} else {
+						log.Printf("\t\t+ %s \n", mem.Email)
+					}
 				}
 				for _, mem := range gr.membersToRemove {
-					log.Printf("\t\t- %s \n", mem.Email)
-					glib.RemoveMember(*clientService, gr.groupToUpdate.Email, mem)
+					err := glib.RemoveMember(*clientService, gr.groupToUpdate.Email, mem)
+					if err != nil {
+						return fmt.Errorf("⚠ Failed to add a member to a group: %v.", err)
+					} else {
+						log.Printf("\t\t- %s \n", mem.Email)
+					}
 				}
 				for _, mem := range gr.membersToUpdate {
-					log.Printf("\t\t~ %s \n", mem.Email)
-					glib.UpdateMembership(*clientService, gr.groupToUpdate.Email, mem)
-
+					err := glib.UpdateMembership(*clientService, gr.groupToUpdate.Email, mem)
+					if err != nil {
+						return fmt.Errorf("⚠ Failed to update membership in a group: %v.", err)
+					} else {
+						log.Printf("\t\t~ %s \n", mem.Email)
+					}
 				}
 			}
 		}
@@ -287,6 +325,7 @@ func SyncMembers(ctx context.Context, clientService *admin.Service, cfgGr *confi
 	var memToUpdate []*config.MemberConfig
 	var memToRemove []*admin.Member
 	currentMembers, _ := glib.GetListOfMembers(clientService, curGr)
+
 	// check members to add
 	for _, member := range cfgGr.Members {
 		foundMem := false
@@ -332,7 +371,7 @@ func SyncOrgUnits(ctx context.Context, clientService *admin.Service, cfg *config
 	// get the current users array
 	currentOus, err := glib.GetListOfOrgUnits(clientService)
 	if err != nil {
-		return fmt.Errorf("⚠ failed to get current org units: %v", err)
+		return err
 	}
 	// config defined users
 	configOus := cfg.OrgUnits
@@ -378,21 +417,30 @@ func SyncOrgUnits(ctx context.Context, clientService *admin.Service, cfg *config
 		if ouToCreate != nil {
 			log.Println("✎ Creating...")
 			for _, ou := range ouToCreate {
-				glib.CreateOrgUnit(*clientService, &ou)
+				err := glib.CreateOrgUnit(*clientService, &ou)
+				if err != nil {
+					return err
+				}
 				log.Printf("\t+ org unit: %s\n", ou.Name)
 			}
 		}
 		if ouToDelete != nil {
 			log.Println("✁ Deleting...")
 			for _, ou := range ouToDelete {
-				glib.DeleteOrgUnit(*clientService, ou)
+				err := glib.DeleteOrgUnit(*clientService, ou)
+				if err != nil {
+					return err
+				}
 				log.Printf("\t- org unit: %s\n", ou.Name)
 			}
 		}
 		if ouToUpdate != nil {
 			log.Println("✎ Updating...")
 			for _, ou := range ouToUpdate {
-				glib.UpdateOrgUnit(*clientService, &ou)
+				err := glib.UpdateOrgUnit(*clientService, &ou)
+				if err != nil {
+					return err
+				}
 				log.Printf("\t~ org unit: %s \n", ou.Name)
 			}
 		}