+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to this module will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [v0.0.4] - 2020-07-28
+## [v0.0.4] - 2020-08-11
 
 - Group configuration expanded by fields: 
-  - ...
+  - Access Type settings 
+	- if is archivied
 - User configuration expanded by fields:
-  - ...
+  - Aliases        
+	- Phones, address 
+	- Recovery Phone & Recovery Email  
+	- Employee & Location information
 
 ## [v0.0.3] - 2020-07-17